split parser and ast to avoid duplicate objects
