+++ src/parse_chunk.hpp
@@ -3,10 +3,7 @@
 
 //#define BOOST_SPIRIT_X3_DEBUG
 #include <boost/spirit/home/x3.hpp>
-#include <boost/spirit/home/x3/support/ast/position_tagged.hpp>
 #include <boost/spirit/home/x3/support/utility/annotate_on_success.hpp>
-#include <boost/fusion/include/adapt_struct.hpp>
-
 
 #include <Rcpp.h>
 
@@ -14,38 +11,7 @@
 #include "parse_option.hpp"
 #include "parser_error_handler.hpp"
 
-namespace client { namespace ast {
-  namespace x3 = boost::spirit::x3;
-
-
-  struct details : x3::position_tagged {
-    std::string name;
-    std::vector<option> options;
-  };
-
-  struct chunk : x3::position_tagged{
-    std::string indent;
-    std::string engine;
-    details d;
-    std::vector<std::string> code;
-  };
-} }
-
-BOOST_FUSION_ADAPT_STRUCT(
-  client::ast::option,
-  name, value
-)
-
-BOOST_FUSION_ADAPT_STRUCT(
-  client::ast::details,
-  name, options
-)
-
-BOOST_FUSION_ADAPT_STRUCT(
-  client::ast::chunk,
-  indent, engine, d, code
-)
-
+#include "parse_chunk_ast.hpp"
 
 namespace client { namespace parser {
   namespace x3 = boost::spirit::x3;
